+++ CHANGES.md
@@ -4,6 +4,8 @@ Unreleased
 - Fix reflection of CHAR columns (#275)
 - Fix reflection of TIMESTAMPTZ columns (#276), thanks to @nvachhar
 - Fix reflection of JSONB columns (#277)
+- Fix compatibility issues with Alembic 1.18 (via SQLA 2.0.47)
+- Update minimum Python version to 3.10
 
 
 # Version 2.0.3